History and Changes in database
Change database datatype for currency from float t decimal. While as history make it independent from the up-to-date inventory

+++ CHKS/Models/Mydb/Cart.cs
@@ -18,7 +18,7 @@ namespace CHKS.Models.mydb
         [Required]
         public int CartId { get; set; }
 
-        public float? Total { get; set; }
+        public decimal? Total { get; set; }
 
         public ICollection<Connector> Connectors { get; set; }
     }
